backend_minerva/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.8.1
attrs==25.4.0
cachetools==6.2.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
defusedxml==0.7.1
Django==5.2.7
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-spectacular==0.29.0
filetype==1.2.0
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth>=2.47.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
inflection==0.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.7
langchain-core==1.2.7
langchain-google-genai==4.2.0
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.4
oauthlib==3.2.2
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.0
pydantic_core==2.41.1
PyJWT==2.9.0
pyparsing==3.2.5
python-decouple==3.8
python3-openid==3.2.0
PyYAML==6.0.3
referencing==0.37.0
reportlab==4.4.4
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.30.0
rsa==4.9.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uuid_utils==0.14.0
xxhash==3.6.0
zstandard==0.25.0
psycopg2-binary==2.9.10
